Got notf to build under windows, no visuals yet

+++ src/common/hash.hpp
@@ -42,7 +42,7 @@ template<typename T, typename... Rest>
 inline void hash_combine(std::size_t& seed, const T& v, Rest&&... rest)
 {
     // see http://stackoverflow.com/a/4948967 for an explanation of the magic number
-    static const size_t magic_number = static_cast<size_t>(powl(2.l, bitsizeof<size_t>()) / ((sqrtl(5.l) + 1) / 2));
+    static const size_t magic_number = static_cast<size_t>(powl(2.l, static_cast<double>(bitsizeof<size_t>())) / ((sqrtl(5.l) + 1) / 2));
     seed ^= std::hash<T>{}(v) + magic_number + (seed << 6) + (seed >> 2);
     hash_combine(seed, std::forward<Rest>(rest)...);
 }
